small bug fix

+++ firmware/shared/src/app/app_fir_filter.c
@@ -37,12 +37,13 @@ float app_fir_filter_apply(AppFIRFilter *filter, float new_sample)
     float result = 0.0f;
     for (uint16_t i = 0; i < filter->window_size; i++)
     {
-        uint16_t sample_index = (filter->index + filter->window_size - i) % filter->window_size;
+        uint16_t sample_index = (uint16_t)((filter->index + filter->window_size - i) % filter->window_size);
         result += filter->samples[sample_index] * filter->coefficients[i];
     }
 
     // Advance the index in the circular buffer
-    filter->index = (filter->index + 1) % filter->window_size;
+    filter->index = (uint16_t)((filter->index + 1) % filter->window_size);
 
     return result;
 }
+
